backend/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-jwt-2fa==0.3.0
Markdown==3.2.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
drf-jwt-2fa==0.3.0
SQLAlchemy==1.3.20
django-url-filter==0.3.5
django_dbconn_retry==0.1.5
pandas==1.1.5